% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplate.R
\name{usetemplate}
\alias{usetemplate}
\title{Create R script from a template}
\usage{
usetemplate(filename, path_dest = ".", path_template = NULL, template = NULL)
}
\arguments{
\item{filename}{A character string of the filename of R script to be created,
without the \code{.R} extension.}

\item{path_dest}{A character string of the destination directory. Defaults to
the current directory.}

\item{path_template}{A character string of the template script directory.
When it is NULL (the default), it will use the built-in template directory.}

\item{template}{A character string of the template filename, without the \code{.R}
extension.}
}
\value{
An R script stored in the \code{path_dest} directory.
}
\description{
Generates an R script from a template
}
\examples{
\dontrun{usetemplate("test", template = "test_template")}
}
